Clean up docs.

+++ codegen/nrbf/tests/binary_array_jagged_offset.rs
@@ -1,6 +1,6 @@
 use std::collections::BTreeMap;
 
-use nrbf::{data_type::Int32, RemotingMessage, Value};
+use nrbf::{RemotingMessage, Value};
 
 #[test]
 fn binary_array_jagged_offset() {
@@ -48,12 +48,12 @@ fn binary_array_jagged_offset() {
 
   let output = RemotingMessage::Value(
     BTreeMap::from_iter([
-      (Int32(1), Value::Array(vec![Value::Ref(Int32(2)), Value::Ref(Int32(3)), Value::Ref(Int32(4))])),
-      (Int32(2), Value::Array(vec![Value::Int32(1)])),
-      (Int32(3), Value::Array(vec![Value::Int32(2), Value::Int32(3)])),
-      (Int32(4), Value::Array(vec![Value::Int32(4), Value::Int32(5), Value::Int32(6)])),
+      (1, Value::Array(vec![Value::Ref(2), Value::Ref(3), Value::Ref(4)])),
+      (2, Value::Array(vec![Value::Int32(1)])),
+      (3, Value::Array(vec![Value::Int32(2), Value::Int32(3)])),
+      (4, Value::Array(vec![Value::Int32(4), Value::Int32(5), Value::Int32(6)])),
     ]),
-    Value::Ref(Int32(1)),
+    Value::Ref(1),
   );
 
   assert_eq!(RemotingMessage::parse(&input), Ok(([].as_slice(), output)));
